+++ Products/MemcachedManager/MemcachedManager.py
@@ -76,10 +76,6 @@ else:
             "ketama": True,
         }
 
-        if memcache.__version__ < "1.2.0":
-            # Recent pylibmc versions (1.2.0 and up) no longer support this option
-            behaviors["cache_lookups"] = True
-
         def __init__(self, servers, debug=False, pickleProtocol=None):
             # pickle protocol is always set to -1 in pylibmc
             self.debug = debug